behavior test base
