+++ modules/javafx.controls/src/test/java/test/javafx/scene/control/TextFieldBehaviorTest.java
@@ -24,12 +24,23 @@
  */
 package test.javafx.scene.control;
 
-import static javafx.scene.input.KeyCode.*;
-import java.util.function.BooleanSupplier;
+import static javafx.scene.input.KeyCode.A;
+import static javafx.scene.input.KeyCode.BACK_SPACE;
+import static javafx.scene.input.KeyCode.C;
+import static javafx.scene.input.KeyCode.DOWN;
+import static javafx.scene.input.KeyCode.END;
+import static javafx.scene.input.KeyCode.ENTER;
+import static javafx.scene.input.KeyCode.ESCAPE;
+import static javafx.scene.input.KeyCode.F1;
+import static javafx.scene.input.KeyCode.HOME;
+import static javafx.scene.input.KeyCode.LEFT;
+import static javafx.scene.input.KeyCode.RIGHT;
+import static javafx.scene.input.KeyCode.SPACE;
+import static javafx.scene.input.KeyCode.V;
+import static javafx.scene.input.KeyCode.X;
 import javafx.event.EventType;
 import javafx.scene.control.IndexRange;
 import javafx.scene.control.TextField;
-import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyEvent;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
@@ -37,53 +48,39 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import com.sun.javafx.tk.Toolkit;
 import test.com.sun.javafx.scene.control.infrastructure.KeyEventFirer;
-import test.com.sun.javafx.scene.control.infrastructure.KeyModifier;
 import test.com.sun.javafx.scene.control.infrastructure.StageLoader;
 
 /**
  * Tests the TextField behavior using public APIs.
  */
-public class TextFieldBehaviorTest {
+public class TextFieldBehaviorTest extends BehaviorTestBase<TextField> {
     private static final EventType<KeyEvent> PRE = KeyEvent.KEY_PRESSED;
     private static final EventType<KeyEvent> TYP = KeyEvent.KEY_TYPED;
     private static final EventType<KeyEvent> REL = KeyEvent.KEY_RELEASED;
     private TextField control;
-    private StageLoader stageLoader;
-    // TODO problem:
-    // KeyEventFirer may not a good idea here because of the way it generates events.
-    // I think we should rather emulate the keyboard, such that the events match those sent by the real thing
-    // i.e. press(SHORTCUT), hit(X), release(SHORTCUT)
-    private KeyEventFirer kb;
-    private int step;
 
     @BeforeEach
-    public void before() {
+    public void beforeEach() {
         control = new TextField();
-        stageLoader = new StageLoader(control);
-        kb = new KeyEventFirer(control);
-        control.requestFocus();
-        Toolkit.getToolkit().firePulse();
+        initStage(control);
     }
 
     @AfterEach
-    public void after() {
-        stageLoader.dispose();
+    public void afterEach() {
+        closeStage();
     }
 
     /**
      * Tests basic typing.
      */
     @Test
-    public void testTyping() {
-        kb.type("hello");
-        check("hello");
-        kb.type(BACK_SPACE, BACK_SPACE, "f");
-        check("helf");
-    }
-
-    private void check(String text) {
-        String s = control.getText();
-        Assertions.assertEquals(text, s);
+    public void testTyping2() {
+        t(
+            "hello",
+            checkText("hello"),
+            BACK_SPACE, BACK_SPACE, "f",
+            checkText("helf")
+        );
     }
 
     /** tests event consumption logic */
@@ -162,8 +159,6 @@ public class TextFieldBehaviorTest {
         );
     }
 
-    // TODO perhaps this should be extracted into a separate class BehaviorTestRunner<T extends Control>
-
     protected Runnable setText(String text) {
         return () -> {
             control.setText(text);
@@ -183,50 +178,5 @@ public class TextFieldBehaviorTest {
             IndexRange expected = new IndexRange(start, end);
             Assertions.assertEquals(expected, v, errorMessage());
         };
-    }
-
-    protected Runnable shift(KeyCode k) {
-        return () -> {
-            kb.keyPressed(k, KeyModifier.SHIFT);
-            kb.keyReleased(k, KeyModifier.SHIFT);
-        };
-    }
-
-    protected Runnable shortcut(KeyCode k) {
-        return () -> {
-            kb.keyPressed(k, KeyModifier.getShortcutKey());
-            kb.keyReleased(k, KeyModifier.getShortcutKey());
-        };
-    }
-
-    protected Runnable x(BooleanSupplier test) {
-        return () -> {
-            boolean result = test.getAsBoolean();
-            Assertions.assertTrue(result, errorMessage());
-        };
-    }
-
-    protected String errorMessage() {
-        return "in step " + step;
-    }
-
-    /**
-     * Executes a test by emulating key press / key releases and various operations upon control.
-     * @param items the sequence of KeyCodes/Runnables
-     */
-    protected void t(Object ... items) {
-        step = 0;
-        for(Object x: items) {
-            if(x instanceof Runnable r) {
-                r.run();
-                Toolkit.getToolkit().firePulse();
-            } else if(x instanceof KeyCode k) {
-                kb.keyPressed(k);
-                kb.keyReleased(k);
-            } else if(x instanceof String s) {
-                kb.type(s);
-            }
-            step++;
-        }
     }
 }
